+++ quantum_inferno/styx_stx.py
@@ -1,12 +1,15 @@
 """"
 Standardized Stockwell transform (stx) with optimization parameters
 After Moukadem et al., 2022, A new optimized Stockwell transform applied on synthetic and real non-stationary signals
-Rederivation in preparation, Garces et al. 2022; last updated 9 May 2022
+Rederivation in preparation, last updated in Garces et al. 2023
+
 """
+from typing import Tuple
+
 import numpy as np
 from scipy.fft import fft, ifft, fftfreq
+
 from quantum_inferno import scales_dyadic as scales
-from typing import Tuple
 from quantum_inferno.utilities.rescaling import is_power_of_two