+++ quantum_inferno/utilities/matrix.py
@@ -1,9 +1,10 @@
 """
 Utilities for matrix operations.
+
 """
 
-import numpy as np
 from enum import Enum
+import numpy as np
 
 
 class MatrixAxis(Enum):